pyproject.toml
[project]
name = "concordia-sim"
version = "0.1.0"
description = "Multi-agent simulation framework built on Concordia v2"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "gdm-concordia>=2.0.0",
    "hydra-core>=1.3.2",
    "hydra-auto-schema>=0.0.5",
    "omegaconf>=2.3.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "commitizen>=3.13.0",
    "bandit>=1.7.7",
    "types-PyYAML>=6.0.0",
]

ollama = [
    "ollama>=0.1.0",
]

all = [
    "concordia-sim[dev,ollama]",
]

[project.scripts]
concordia-sim = "run_experiment:main"

[project.urls]
Homepage = "https://github.com/yourusername/concordia-sim"
Documentation = "https://github.com/yourusername/concordia-sim#readme"
Repository = "https://github.com/yourusername/concordia-sim"
Issues = "https://github.com/yourusername/concordia-sim/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "scenarios"]

[tool.ruff]
target-version = "py312"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
    "ARG002", # unused method argument (often required for interface compliance)
    "TCH002", # move third-party import into type-checking block
    "TCH003", # move standard library import into type-checking block
    "B027",   # empty method in abstract base class (optional hooks)
    "UP040",  # TypeAlias annotation (Python 3.12 syntax not required)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["ARG", "S101", "E402"]
"run_experiment.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["src", "scenarios"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
plugins = []

[[tool.mypy.overrides]]
module = [
    "concordia.*",
    "sentence_transformers.*",
    "hydra.*",
    "omegaconf.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src", "scenarios"]
branch = true
omit = [
    "tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 70

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "pyproject.toml:version",
]
tag_format = "v$version"
update_changelog_on_bump = true

[tool.bandit]
skips = ["B311"]  # random is fine for simulation, not used for security
